server/requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.14.3
bleach==6.3.0
braceexpand==0.1.7
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
ctc-segmentation==1.7.4
cuda-bindings==13.1.0
cuda-core==0.4.2
cuda-pathfinder==1.3.3
cuda-python==13.1.0
cycler==0.12.1
cython==3.2.2
cytoolz==1.1.0
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
docopt==0.6.2
docstring-parser==0.17.0
editdistance==0.8.1
einops==0.8.1
eval-type-backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.1
fastapi==0.124.2
fastjsonschema==2.21.2
fiddle==0.3.0
filelock==3.19.1
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
gitpython==3.1.45
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib-metadata==8.7.0
indic-numtowords==1.1.0
inflect==7.5.0
intervaltree==3.1.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
jiwer==3.1.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kaldi-python-io==1.2.2
kaldialign==0.9.1
kiwisolver==1.4.9
lark==1.3.1
lazy-loader==0.4
lhotse==1.32.1
libcst==1.8.6
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
livekit==1.0.20
livekit-agents==1.3.6
livekit-api==1.1.0
livekit-blingfire==1.0.0
livekit-protocol==1.1.1
llvmlite==0.45.1
mako==1.3.10
markdown==3.10
markdown-it-py==4.0.0
markupsafe==2.1.5
marshmallow==4.1.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.1.6
minio==7.2.10
mistune==3.1.4
ml-dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@66ffb38338aa03cf938cb17b53c5dbb509bc17fe
nest-asyncio==1.6.0
networkx==3.3
notebook==7.5.0
notebook-shim==0.2.4
num2words==0.5.14
numba==0.62.1
numba-cuda==0.22.1
numexpr==2.13.1
numpy==1.26.4
nv-one-logger-core==2.3.1
nv-one-logger-pytorch-lightning-integration==2.3.1
nv-one-logger-training-telemetry==2.3.1
nvidia-cuda-cccl==13.1.78
nvidia-cuda-nvrtc==13.1.80
nvidia-cuda-runtime==13.1.80
nvidia-nvjitlink==13.1.80
nvidia-nvvm==13.1.80
omegaconf==2.3.0
onnx==1.20.0
openai==2.9.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
optuna==4.6.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.3
pure-eval==0.2.3
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pyjwt==2.10.1
pyloudnorm==0.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytorch-lightning==2.6.0
pytz==2025.2
pywinpty==3.0.2
pyyaml==6.0.3
pyzmq==27.1.0
rapidfuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.15
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
send2trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
sox==1.5.0
soxr==1.0.0
sqlalchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
strenum==0.4.15
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.3.0
toolz==1.1.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typeguard==4.4.4
typer==0.20.0
types-protobuf==6.32.1.20251105
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.6.0
uvicorn==0.38.0
wandb==0.23.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.4
wget==3.2
whisper-normalizer==0.1.12
widgetsnbextension==4.0.15
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
